Move ChatThread back to components folder
