+++ change/@azure-communication-ui-ddcede31-ec86-4156-97be-94d9f593e21f.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "Refactor ChatThread component and optimize ReadReceipt data flow",
+  "packageName": "@azure/communication-ui",
+  "email": "easony@microsoft.com",
+  "dependentChangeType": "patch"
+}
