+++ packages/communication-ui/src/components/index.ts
@@ -19,6 +19,8 @@ export {
 export { SendBoxComponent } from './SendBox';
 export { ReadReceiptComponent } from './ReadReceipt';
 export type { ReadReceiptProps } from './ReadReceipt';
+export { ChatThreadComponent, ChatThread } from './ChatThread';
+export type { ChatThreadProps, ChatThreadStylesProps } from './ChatThread';
 export { StreamMedia } from './StreamMedia';
 export { ParticipantItem } from './ParticipantItem';
 export {
