+++ packages/communication-ui/src/components/styles/ChatThread.styles.ts
